# Conflicts:
#	common/src/main/java/org/openbaton/nfvo/common/utils/viminstance/VimInstanceUtils.java
#	dashboard

+++ vnfm/vnfm-impl/src/main/java/org/openbaton/nfvo/vnfm_reg/VnfmManager.java
@@ -55,13 +55,17 @@ import org.openbaton.catalogue.nfvo.messages.OrVnfmUpdateMessage;
 import org.openbaton.catalogue.nfvo.viminstances.BaseVimInstance;
 import org.openbaton.exceptions.BadFormatException;
 import org.openbaton.exceptions.NotFoundException;
+import org.openbaton.exceptions.PluginException;
+import org.openbaton.exceptions.VimException;
 import org.openbaton.nfvo.common.internal.model.EventFinishNFVO;
 import org.openbaton.nfvo.common.internal.model.EventNFVO;
+import org.openbaton.nfvo.core.interfaces.VimManagement;
 import org.openbaton.nfvo.repositories.NetworkServiceDescriptorRepository;
 import org.openbaton.nfvo.repositories.NetworkServiceRecordRepository;
 import org.openbaton.nfvo.repositories.VNFDRepository;
 import org.openbaton.nfvo.repositories.VNFRRepository;
 import org.openbaton.nfvo.repositories.VimRepository;
+import org.openbaton.nfvo.repositories.VirtualLinkRecordRepository;
 import org.openbaton.nfvo.repositories.VnfPackageRepository;
 import org.openbaton.vnfm.interfaces.manager.MessageGenerator;
 import org.openbaton.vnfm.interfaces.sender.VnfmSender;
@@ -104,9 +108,14 @@ public class VnfmManager
   @Autowired private VNFRRepository vnfrRepository;
   @Autowired private VimRepository vimInstanceRepository;
   @Autowired private MessageGenerator generator;
+  @Autowired private VimManagement vimManagement;
+  @Autowired private VirtualLinkRecordRepository vlrRepository;
 
   @Value("${nfvo.start.ordered:false}")
   private boolean ordered;
+
+  @Value("${nfvo.networks.dedicated:false}")
+  private boolean dedicatedNetworks;
 
   private static <K, V extends Comparable<? super V>> Map<K, V> sortByValue(Map<K, V> map) {
     List<Entry<K, V>> list = new LinkedList<>(map.entrySet());
@@ -252,6 +261,13 @@ public class VnfmManager
             status = vnfr.getStatus();
           }
         }
+        status =
+            networkServiceRecord
+                    .getVnfr()
+                    .stream()
+                    .anyMatch(vnfr -> vnfr.getStatus().ordinal() == Status.TERMINATED.ordinal())
+                ? Status.TERMINATED
+                : status;
 
         log.debug("Setting NSR status to: " + status);
         networkServiceRecord.setStatus(status);
@@ -314,20 +330,45 @@ public class VnfmManager
             }
           } catch (OptimisticLockingFailureException e) {
             log.error(
-                "FFFF OptimisticLockingFailureException while setting the task of the NSR. Don't worry we will try it"
-                    + " again.");
+                "Got OptimisticLockingFailureException while setting the task of the NSR. Don't worry I'll try again.");
             savedNsr = false;
           }
         } else {
           log.debug("Nsr is ACTIVE but not all vnfr have been received");
         }
       } while (!savedNsr);
-    } else if (status.ordinal() == Status.TERMINATED.ordinal()) {
+    } else if (status.ordinal() == Status.TERMINATED.ordinal()
+        && networkServiceRecord
+            .getVnfr()
+            .stream()
+            .allMatch(vnfr -> vnfr.getStatus().ordinal() == Status.TERMINATED.ordinal())) {
       publishEvent(
           Action.RELEASE_RESOURCES_FINISH,
           networkServiceRecord,
           networkServiceRecord.getProjectId());
       nsrRepository.delete(networkServiceRecord);
+      if (dedicatedNetworks) {
+
+        networkServiceRecord
+            .getVlr()
+            .parallelStream()
+            .filter(
+                virtualLinkRecord ->
+                    vlrRepository.findByExtId(virtualLinkRecord.getExtId()).size() == 0)
+            .forEach(
+                vlr -> {
+                  try {
+                    log.info(
+                        String.format("Deleting Network: %s [%s]", vlr.getName(), vlr.getExtId()));
+                    vimManagement.deleteNetwork(vlr);
+                  } catch (PluginException | VimException e) {
+                    e.printStackTrace();
+                    log.error(String.format("Not Able to delete network %s!", vlr.getExtId()));
+                  } catch (NotFoundException e) {
+                    log.warn(String.format("%s. probably already deleted...", e.getMessage()));
+                  }
+                });
+      }
     }
 
     log.trace("Thread: " + Thread.currentThread().getId() + " finished findAndSet");
